+++ MapboxNavigation/FeedbackViewController.swift
@@ -3,29 +3,47 @@ import MapboxCoreNavigation
 import AVFoundation
 
 extension FeedbackViewController: UIViewControllerTransitioningDelegate {
-    func animationController(forDismissed dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+    @objc public func animationController(forDismissed dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
         abortAutodismiss()
         return DismissAnimator()
     }
     
-    func animationController(forPresented presented: UIViewController, presenting: UIViewController, source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+    @objc public func animationController(forPresented presented: UIViewController, presenting: UIViewController, source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
         return PresentAnimator()
     }
     
-    func interactionControllerForDismissal(using animator: UIViewControllerAnimatedTransitioning) -> UIViewControllerInteractiveTransitioning? {
+    @objc public func interactionControllerForDismissal(using animator: UIViewControllerAnimatedTransitioning) -> UIViewControllerInteractiveTransitioning? {
         return interactor.hasStarted ? interactor : nil
     }
 }
 
-typealias FeedbackSection = [FeedbackItem]
+/**
+ The `FeedbackViewControllerDelegate` protocol provides methods for responding to feedback events.
+ */
+@objc public protocol FeedbackViewControllerDelegate {
+    
+    /**
+     Called when the user opens the feedback form.
+     */
+    @objc optional func  feedbackViewControllerDidOpen(_ feedbackViewController: FeedbackViewController)
+    
+    /**
+     Called when the user submits a feedback event.
+     */
+    @objc(feedbackViewController:didSendFeedbackItem:UUID:)
+    optional func feedbackViewController(_ feedbackViewController: FeedbackViewController, didSend feedbackItem: FeedbackItem, uuid: UUID)
+    
+    /**
+     Called when a `FeedbackViewController` is dismissed for any reason without giving explicit feedback.
+     */
+    @objc optional func feedbackViewControllerDidCancel(_ feedbackViewController: FeedbackViewController)
+}
 
-class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecognizerDelegate {
-    
-    typealias SendFeedbackHandler = (FeedbackItem) -> Void
-    
-    var sendFeedbackHandler: SendFeedbackHandler?
-    var dismissFeedbackHandler: (() -> Void)?
-    var sections = [FeedbackSection]()
+/**
+ A view controller containing a grid of buttons the user can use to denote an issue their current navigation experience.
+ */
+@objc(MBFeedbackViewController)
+public class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecognizerDelegate {
     var activeFeedbackItem: FeedbackItem?
     
     static let sceneTitle = NSLocalizedString("FEEDBACK_TITLE", value: "Report Problem", comment: "Title of view controller for sending feedback")
@@ -34,6 +52,13 @@ class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecog
     static let verticalCellPadding: CGFloat = 20.0
     
     let interactor = Interactor()
+    
+    /**
+     The feedback items that are visible and selectable by the user.
+     */
+    public var sections: [FeedbackItem] = [.turnNotAllowed, .closure, .reportTraffic, .confusingInstructions, .generalMapError, .badRoute]
+    
+    @objc public weak var delegate: FeedbackViewControllerDelegate?
     
     lazy var collectionView: UICollectionView = {
         let view: UICollectionView = UICollectionView(frame: .zero, collectionViewLayout: flowLayout)
@@ -71,24 +96,58 @@ class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecog
         return fullHeight
     }
     
-    override func viewDidLoad() {
+    /**
+     The events manager used to send feedback events.
+     */
+    public var eventsManager: EventsManager
+    
+    var uuid: UUID {
+        return eventsManager.recordFeedback()
+    }
+    
+    /**
+     Initialize a new FeedbackViewController from an `EventsManager`.
+     */
+    @objc public init(eventsManager: EventsManager) {
+        self.eventsManager = eventsManager
+        super.init(nibName: nil, bundle: nil)
+        commonInit()
+    }
+
+    public override func encode(with aCoder: NSCoder) {
+        aCoder.encode(eventsManager, forKey: "EventsManager")
+    }
+    
+    required public init?(coder aDecoder: NSCoder) {
+        eventsManager = aDecoder.decodeObject(of: [EventsManager.self], forKey: "EventsManager") as? EventsManager ?? EventsManager(accessToken: nil)
+        super.init(coder: aDecoder)
+        commonInit()
+    }
+    
+    func commonInit() {
+        self.modalPresentationStyle = .custom
+        self.transitioningDelegate = self
+    }
+    
+    override public func viewDidLoad() {
         super.viewDidLoad()
         setupViews()
         setupConstraints()
         view.layoutIfNeeded()
         transitioningDelegate = self
         view.backgroundColor = .white
-        progressBar.barColor = #colorLiteral(red: 0.9347146749, green: 0.5047877431, blue: 0.1419634521, alpha: 1)
         enableDraggableDismiss()
     }
     
-    override func viewWillAppear(_ animated: Bool) {
+    override public func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         progressBar.progress = 1
     }
     
-    override func viewDidAppear(_ animated: Bool) {
+    override public func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
+        
+        delegate?.feedbackViewControllerDidOpen?(self)
         
         UIView.animate(withDuration: FeedbackViewController.autoDismissInterval) {
             self.progressBar.progress = 0
@@ -97,7 +156,7 @@ class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecog
         enableAutoDismiss()
     }
     
-    override func willTransition(to newCollection: UITraitCollection, with coordinator: UIViewControllerTransitionCoordinator) {
+    override public func willTransition(to newCollection: UITraitCollection, with coordinator: UIViewControllerTransitionCoordinator) {
         super.willTransition(to: newCollection, with: coordinator)
         
         // Dismiss the feedback view when switching between landscape and portrait mode.
@@ -126,12 +185,15 @@ class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecog
         NSObject.cancelPreviousPerformRequests(withTarget: self, selector: #selector(dismissFeedback), object: nil)
     }
     
-    @objc func dismissFeedback() {
+    /**
+     Instantly dismisses the FeedbackViewController if it is currently presented.
+     */
+    @objc public func dismissFeedback() {
         abortAutodismiss()
-        dismissFeedbackHandler?()
+        dismissFeedbackItem()
     }
     
-    func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldReceive touch: UITouch) -> Bool {
+    public func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldReceive touch: UITouch) -> Bool {
         // Only respond to touches outside/behind the view
         let isDescendant = touch.view?.isDescendant(of: view) ?? true
         return !isDescendant
@@ -165,12 +227,32 @@ class FeedbackViewController: UIViewController, DismissDraggable, UIGestureRecog
         progressBar.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
         progressBar.bottomAnchor.constraint(equalTo: view.safeBottomAnchor).isActive = true
     }
+    
+    func send(_ item: FeedbackItem) {
+        delegate?.feedbackViewController?(self, didSend: item, uuid: uuid)
+        eventsManager.updateFeedback(uuid: uuid, type: item.feedbackType, source: .user, description: nil)
+        
+        guard let parent = presentingViewController else {
+            dismiss(animated: true)
+            return
+        }
+        
+        dismiss(animated: true) {
+            DialogViewController().present(on: parent)
+        }
+    }
+    
+    func dismissFeedbackItem() {
+        delegate?.feedbackViewControllerDidCancel?(self)
+        eventsManager.cancelFeedback(uuid: uuid)
+        dismiss(animated: true, completion: nil)
+    }
 }
 
 extension FeedbackViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+    @objc public func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: FeedbackCollectionViewCell.defaultIdentifier, for: indexPath) as! FeedbackCollectionViewCell
-        let item = sections[indexPath.section][indexPath.row]
+        let item = sections[indexPath.row]
         
         cell.titleLabel.text = item.title
         cell.imageView.tintColor = .clear
@@ -179,15 +261,15 @@ extension FeedbackViewController: UICollectionViewDataSource {
         return cell
     }
     
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
+    @objc public func numberOfSections(in collectionView: UICollectionView) -> Int {
+        return 1
+    }
+    
+    @objc public func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return sections.count
     }
     
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return sections[section].count
-    }
-    
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
+    @objc public func scrollViewDidScroll(_ scrollView: UIScrollView) {
         // In case the view is scrolled, dismiss the feedback window immediately
         // and reset the `progressBar` back to a full progress.
         abortAutodismiss()
@@ -196,24 +278,22 @@ extension FeedbackViewController: UICollectionViewDataSource {
 }
 
 extension FeedbackViewController: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+    @objc public func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         abortAutodismiss()
-        let item = sections[indexPath.section][indexPath.row]
-        sendFeedbackHandler?(item)
+        let item = sections[indexPath.row]
+        send(item)
     }
 }
 
 extension FeedbackViewController: UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+    @objc public func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         let availableWidth = collectionView.bounds.width
         // 3 columns and 2 rows in portrait mode.
         // 6 columns and 1 row in landscape mode.
-        let items = sections[indexPath.section]
         let width = traitCollection.verticalSizeClass == .compact
-            ? floor(availableWidth / CGFloat(items.count))
-            : floor(availableWidth / CGFloat(items.count / 2))
-        let item = sections[indexPath.section][indexPath.row]
+            ? floor(availableWidth / CGFloat(sections.count))
+            : floor(availableWidth / CGFloat(sections.count / 2))
+        let item = sections[indexPath.row]
         let titleHeight = item.title.height(constrainedTo: width, font: FeedbackCollectionViewCell.Constants.titleFont)
         let cellHeight: CGFloat = FeedbackCollectionViewCell.Constants.imageSize.height
                                   + FeedbackCollectionViewCell.Constants.padding